requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.7
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
backoff==2.2.1
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.2
datasets==3.0.1
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dspy-ai==2.5.0
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
litellm==1.48.4
magicattr==0.1.6
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.1.0
multiprocess==0.70.16
numpy==2.1.1
openai==1.50.2
optuna==4.0.0
packaging==24.1
pandas==2.2.3
pyarrow==17.0.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rpds-py==0.20.0
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
structlog==24.4.0
tiktoken==0.7.0
tokenizers==0.20.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==2.2.3
wheel==0.44.0
xxhash==3.5.0
yarl==1.13.1
zipp==3.20.2
